Create separate backend helpers
